backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
PyJWT
